minor cleanup things
